+++ internal/form/photo.go
@@ -45,8 +45,8 @@ type Photo struct {
 	PhotoScan        bool      `json:"Scan"`
 	PhotoPanorama    bool      `json:"Panorama"`
 	PhotoAltitude    int       `json:"Altitude"`
-	PhotoLat         float32   `json:"Lat"`
-	PhotoLng         float32   `json:"Lng"`
+	PhotoLat         float64   `json:"Lat"`
+	PhotoLng         float64   `json:"Lng"`
 	PhotoIso         int       `json:"Iso"`
 	PhotoFocalLength int       `json:"FocalLength"`
 	PhotoFNumber     float32   `json:"FNumber"`
